+++ Sources/Litext/LTXLabel/Interaction/LTXLabel+Interaction.swift
@@ -14,6 +14,7 @@ extension LTXLabel {
     func setInteractionStateToBegin(initialLocation: CGPoint) {
         interactionState.initialTouchLocation = initialLocation
         interactionState.isFirstMove = true
+        isInteractionInProgress = true
     }
 
     func bumpClickCountIfWithinTimeGap() {
